feat(rag): extract generation stack into src/rag/ package
Split config loaders, LLM providers, and generate/stream entrypoints so
thin FastAPI can import a package without Streamlit coupling. Public
imports stay `from rag import ...`.
